Cryptlib: move time definitions to their own header
This moves the definitions of time-related types and structs to their
own header so we can include it elsewhere.
